Refactor module list client into modular structure
