+++ crates/core/src/client/module_list/parsing.rs
@@ -1,0 +1,241 @@
+use super::super::{ClientError, FEATURE_UNAVAILABLE_EXIT_CODE, daemon_error};
+use super::DaemonAddress;
+
+pub(crate) fn strip_prefix_ignore_ascii_case<'a>(text: &'a str, prefix: &str) -> Option<&'a str> {
+    if text.len() < prefix.len() {
+        return None;
+    }
+
+    let (candidate, remainder) = text.split_at(prefix.len());
+    if candidate.eq_ignore_ascii_case(prefix) {
+        Some(remainder)
+    } else {
+        None
+    }
+}
+
+pub(crate) fn parse_bracketed_host(
+    host: &str,
+    default_port: u16,
+) -> Result<(String, u16), ClientError> {
+    let (addr, remainder) = host.split_once(']').ok_or_else(|| {
+        daemon_error(
+            "invalid bracketed daemon host",
+            FEATURE_UNAVAILABLE_EXIT_CODE,
+        )
+    })?;
+
+    let decoded = decode_host_component(addr)?;
+
+    if remainder.is_empty() {
+        return Ok((decoded, default_port));
+    }
+
+    let port = remainder
+        .strip_prefix(':')
+        .ok_or_else(|| {
+            daemon_error(
+                "invalid bracketed daemon host",
+                FEATURE_UNAVAILABLE_EXIT_CODE,
+            )
+        })?
+        .parse::<u16>()
+        .map_err(|_| daemon_error("invalid daemon port", FEATURE_UNAVAILABLE_EXIT_CODE))?;
+
+    Ok((decoded, port))
+}
+
+pub(crate) fn decode_host_component(input: &str) -> Result<String, ClientError> {
+    decode_percent_component(
+        input,
+        invalid_percent_encoding_error,
+        invalid_host_utf8_error,
+    )
+}
+
+pub(crate) fn decode_daemon_username(input: &str) -> Result<String, ClientError> {
+    decode_percent_component(
+        input,
+        invalid_username_percent_encoding_error,
+        invalid_username_utf8_error,
+    )
+}
+
+pub(crate) fn decode_percent_component(
+    input: &str,
+    invalid_percent: fn() -> ClientError,
+    invalid_utf8: fn() -> ClientError,
+) -> Result<String, ClientError> {
+    if !input.contains('%') {
+        return Ok(input.to_string());
+    }
+
+    let mut decoded = Vec::with_capacity(input.len());
+    let bytes = input.as_bytes();
+    let mut index = 0;
+
+    while index < bytes.len() {
+        if bytes[index] == b'%' {
+            if index + 2 >= bytes.len() {
+                return Err(invalid_percent());
+            }
+
+            let hi = hex_value(bytes[index + 1]).ok_or_else(invalid_percent)?;
+            let lo = hex_value(bytes[index + 2]).ok_or_else(invalid_percent)?;
+
+            decoded.push((hi << 4) | lo);
+            index += 3;
+            continue;
+        }
+
+        decoded.push(bytes[index]);
+        index += 1;
+    }
+
+    String::from_utf8(decoded).map_err(|_| invalid_utf8())
+}
+
+pub(crate) fn hex_value(byte: u8) -> Option<u8> {
+    match byte {
+        b'0'..=b'9' => Some(byte - b'0'),
+        b'a'..=b'f' => Some(byte - b'a' + 10),
+        b'A'..=b'F' => Some(byte - b'A' + 10),
+        _ => None,
+    }
+}
+
+pub(crate) fn invalid_percent_encoding_error() -> ClientError {
+    daemon_error(
+        "invalid percent-encoding in daemon URL",
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+    )
+}
+
+pub(crate) fn invalid_host_utf8_error() -> ClientError {
+    daemon_error(
+        "daemon host contains invalid UTF-8",
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+    )
+}
+
+pub(crate) fn invalid_username_percent_encoding_error() -> ClientError {
+    daemon_error(
+        "invalid percent-encoding in daemon username",
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+    )
+}
+
+pub(crate) fn invalid_username_utf8_error() -> ClientError {
+    daemon_error(
+        "daemon username contains invalid UTF-8",
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+    )
+}
+
+pub(crate) fn split_host_port(input: &str) -> Option<(&str, &str)> {
+    let idx = input.rfind(':')?;
+    Some((&input[..idx], &input[idx + 1..]))
+}
+
+pub(crate) fn split_daemon_username(input: &str) -> Result<(Option<&str>, &str), ClientError> {
+    if let Some((username, remainder)) = input.split_once('@') {
+        if username.is_empty() {
+            return Err(daemon_error(
+                "daemon username must be non-empty",
+                FEATURE_UNAVAILABLE_EXIT_CODE,
+            ));
+        }
+        Ok((Some(username), remainder))
+    } else {
+        Ok((None, input))
+    }
+}
+
+pub(crate) fn split_daemon_host_module(input: &str) -> Result<Option<(&str, &str)>, ClientError> {
+    if !input.contains('[') {
+        let segments = input.split("::");
+        if segments.clone().count() > 2 {
+            return Err(daemon_error(
+                "IPv6 daemon addresses must be enclosed in brackets",
+                FEATURE_UNAVAILABLE_EXIT_CODE,
+            ));
+        }
+    }
+
+    let mut in_brackets = false;
+    let mut previous_colon = None;
+
+    for (idx, ch) in input.char_indices() {
+        match ch {
+            '[' => {
+                in_brackets = true;
+                previous_colon = None;
+            }
+            ']' => {
+                in_brackets = false;
+                previous_colon = None;
+            }
+            ':' if !in_brackets => {
+                if let Some(prev) = previous_colon.filter(|prev| *prev + 1 == idx) {
+                    let host = &input[..prev];
+                    if !host.contains('[') {
+                        let colon_count = host.chars().filter(|&ch| ch == ':').count();
+                        if colon_count > 1 {
+                            return Err(daemon_error(
+                                "IPv6 daemon addresses must be enclosed in brackets",
+                                FEATURE_UNAVAILABLE_EXIT_CODE,
+                            ));
+                        }
+                    }
+                    let module = &input[idx + 1..];
+                    return Ok(Some((host, module)));
+                }
+                previous_colon = Some(idx);
+            }
+            _ => {
+                previous_colon = None;
+            }
+        }
+    }
+
+    Ok(None)
+}
+
+pub(crate) struct ParsedDaemonTarget {
+    pub(crate) address: DaemonAddress,
+    pub(crate) username: Option<String>,
+}
+
+pub(crate) fn parse_host_port(
+    input: &str,
+    default_port: u16,
+) -> Result<ParsedDaemonTarget, ClientError> {
+    const DEFAULT_HOST: &str = "localhost";
+
+    let (username, input) = split_daemon_username(input)?;
+    let username = username.map(decode_daemon_username).transpose()?;
+
+    if input.is_empty() {
+        let address = DaemonAddress::new(DEFAULT_HOST.to_string(), default_port)?;
+        return Ok(ParsedDaemonTarget { address, username });
+    }
+
+    if let Some(host) = input.strip_prefix('[') {
+        let (address, port) = parse_bracketed_host(host, default_port)?;
+        let address = DaemonAddress::new(address, port)?;
+        return Ok(ParsedDaemonTarget { address, username });
+    }
+
+    if let Some((host, port)) = split_host_port(input) {
+        let port = port
+            .parse::<u16>()
+            .map_err(|_| daemon_error("invalid daemon port", FEATURE_UNAVAILABLE_EXIT_CODE))?;
+        let host = decode_host_component(host)?;
+        let address = DaemonAddress::new(host, port)?;
+        return Ok(ParsedDaemonTarget { address, username });
+    }
+
+    let host = decode_host_component(input)?;
+    let address = DaemonAddress::new(host, default_port)?;
+    Ok(ParsedDaemonTarget { address, username })
+}
