+++ crates/core/src/client/module_list/types.rs
@@ -1,0 +1,61 @@
+use std::fmt;
+
+use super::super::{ClientError, FEATURE_UNAVAILABLE_EXIT_CODE, daemon_error};
+
+/// Target daemon address used for module listing requests.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct DaemonAddress {
+    pub(crate) host: String,
+    pub(crate) port: u16,
+}
+
+impl DaemonAddress {
+    /// Creates a new daemon address from the supplied host and port.
+    pub fn new(host: String, port: u16) -> Result<Self, ClientError> {
+        let trimmed = host.trim();
+        if trimmed.is_empty() {
+            return Err(daemon_error(
+                "daemon host must be non-empty",
+                FEATURE_UNAVAILABLE_EXIT_CODE,
+            ));
+        }
+        Ok(Self {
+            host: trimmed.to_string(),
+            port,
+        })
+    }
+
+    /// Returns the daemon host name or address.
+    #[must_use]
+    pub fn host(&self) -> &str {
+        &self.host
+    }
+
+    /// Returns the daemon TCP port.
+    #[must_use]
+    pub const fn port(&self) -> u16 {
+        self.port
+    }
+
+    pub(crate) fn socket_addr_display(&self) -> SocketAddrDisplay<'_> {
+        SocketAddrDisplay {
+            host: &self.host,
+            port: self.port,
+        }
+    }
+}
+
+pub(crate) struct SocketAddrDisplay<'a> {
+    pub(super) host: &'a str,
+    pub(super) port: u16,
+}
+
+impl fmt::Display for SocketAddrDisplay<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        if self.host.contains(':') && !self.host.starts_with('[') {
+            write!(f, "[{}]:{}", self.host, self.port)
+        } else {
+            write!(f, "{}:{}", self.host, self.port)
+        }
+    }
+}
